+++ db/model/coach.go
@@ -23,7 +23,7 @@ type CoachModel struct {
 	YearsOfWork                  string `json:"years_of_work"`             //从业时长
 	TotalCompleteLesson          string `json:"total_complete_lesson"`     //累计上课节数
 	CanShow                      int    `json:"can_show"`                  //是否可以在平台展示，0=可展示 1=不可展示
-	CloneLessonUnAvaliableSwitch int    `json:"clone_lesson_unava_switch"` //是否把可约状态同步到下周 0=默认同步
+	CloneLessonUnAvaliableSwitch uint8  `json:"clone_lesson_unava_switch"` //是否把可约状态同步到下周 0=默认同步
 
 }
